--- scripts/examples/example_dist_generated_10.json
@@ -1,102 +1,0 @@
-{
-  "totalAmount": "47674000000000",
-  "merkleRoot": "0x971af7943ef35c1951d2b34a0f543aceea4d990b43386fdc0fc61bfc50644c05",
-  "claims": {
-    "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC": {
-      "beneficiary": "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC",
-      "amount": 1000000000000,
-      "proof": [
-        "0x753e810ed621eaa35348ddb2bbcf81e93828fd6a69b15029589ea31c093c99ef",
-        "0x4662c3b26baece43e1444c93540566b8a9ed3e88b259b63970091be98c8fbcc8",
-        "0x07f117b1d7aa852c979367c249b57b7c1cb98dfeccb06ac0ee719d6a07880d99",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    },
-    "0x005511deD6Aa2012DaE1307290779ED52639cCee": {
-      "beneficiary": "0x005511deD6Aa2012DaE1307290779ED52639cCee",
-      "amount": 2000000000000,
-      "proof": [
-        "0x81c73e315412a576f33b7adfae9239e98aefe50869ec606c84b1c2dc000d7484",
-        "0x4662c3b26baece43e1444c93540566b8a9ed3e88b259b63970091be98c8fbcc8",
-        "0x07f117b1d7aa852c979367c249b57b7c1cb98dfeccb06ac0ee719d6a07880d99",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    },
-    "0x9eeE3a95f91b23f2bC186B5977E2fC9d4636a229": {
-      "beneficiary": "0x9eeE3a95f91b23f2bC186B5977E2fC9d4636a229",
-      "amount": 9360000000000,
-      "proof": [
-        "0xa5eb23d28d8838b8342181e734a113680afa3ee2d50c25c45a40bcccd8764207",
-        "0x2a06099e1436f5ec2b724ad6a3fd99c2f0176c538d36fd085eb029eee468443f",
-        "0x07f117b1d7aa852c979367c249b57b7c1cb98dfeccb06ac0ee719d6a07880d99",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    },
-    "0x61b4b98460cd58BCc3758E6885AE4aaAC7f213dE": {
-      "beneficiary": "0x61b4b98460cd58BCc3758E6885AE4aaAC7f213dE",
-      "amount": 4230000000000,
-      "proof": [
-        "0x64e7bd8c9cb823e29dfbc07282c8616ac67d26aebbe3e69946eec626511cfa6a",
-        "0x363ac2361beaa633d6ef504ff1a4336b8cf2658c34c9086ea23e22bb7ef3d9f5",
-        "0xea4c5d91ad89102ba8b7e93436081b349452ac401ad6ab14d1036408c188a507",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    },
-    "0x38445383F364BC835BDAa25e09ab8F1F18479CaE": {
-      "beneficiary": "0x38445383F364BC835BDAa25e09ab8F1F18479CaE",
-      "amount": 6600000000000,
-      "proof": [
-        "0x59247719c52555d9f6311de65206f28c6da50da96b1c87b8e61ba85d3e01df6b",
-        "0x53253b79a3e601d8ac3ff09e7d978b889e5fbc9d70ddd08a3f2238a30d7a4e90",
-        "0xea4c5d91ad89102ba8b7e93436081b349452ac401ad6ab14d1036408c188a507",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    },
-    "0x0C51419B21c884388De6118042f0f9E5b5296A32": {
-      "beneficiary": "0x0C51419B21c884388De6118042f0f9E5b5296A32",
-      "amount": 7770000000000,
-      "proof": [
-        "0x3abf4130c0df53bef24c0e328b7c74b900735079684c787a440708392430f0f4",
-        "0x53253b79a3e601d8ac3ff09e7d978b889e5fbc9d70ddd08a3f2238a30d7a4e90",
-        "0xea4c5d91ad89102ba8b7e93436081b349452ac401ad6ab14d1036408c188a507",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    },
-    "0x8324e107Ab17918Ec3FCE05dC581913223762eb5": {
-      "beneficiary": "0x8324e107Ab17918Ec3FCE05dC581913223762eb5",
-      "amount": 4340000000000,
-      "proof": [
-        "0x8cc91419c608318d3cd299e84fe2541536ea63146cdc216e0023f05a63b6e3f0",
-        "0x2a06099e1436f5ec2b724ad6a3fd99c2f0176c538d36fd085eb029eee468443f",
-        "0x07f117b1d7aa852c979367c249b57b7c1cb98dfeccb06ac0ee719d6a07880d99",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    },
-    "0x729EE12F085cf02a7c365f81192f7774572663cA": {
-      "beneficiary": "0xecd958EF8F5Dc16Cec4ef7d5441bFd9163C3f725",
-      "amount": 8840000000000,
-      "proof": [
-        "0x6993947da2248f7bf046f9ab27b7cc4f2625db627cf163b7e1a31b0018afc26d",
-        "0x363ac2361beaa633d6ef504ff1a4336b8cf2658c34c9086ea23e22bb7ef3d9f5",
-        "0xea4c5d91ad89102ba8b7e93436081b349452ac401ad6ab14d1036408c188a507",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    },
-    "0x91aa9159242c298810b280E3AfFC0b4B289b60cc": {
-      "beneficiary": "0xecd958EF8F5Dc16Cec4ef7d5441bFd9163C3f725",
-      "amount": 3000000000000,
-      "proof": [
-        "0xa659c2d8eecf38adbe9c83f488e2a823a8deddd3103e20e74c923f00275a0d19",
-        "0x2406b5bd435b4c58e6f25507f0d5766eea7777e07a5361e2824446954e1e1213"
-      ]
-    },
-    "0xE4c8d3bcf8C87D73CE38Ab2DC288d309072ee4E7": {
-      "beneficiary": "0xecd958EF8F5Dc16Cec4ef7d5441bFd9163C3f725",
-      "amount": 534000000000,
-      "proof": [
-        "0xf52211cbab67fddf51e81a0461685bc441bd353802343c74924f9f0788de9cdb",
-        "0x2406b5bd435b4c58e6f25507f0d5766eea7777e07a5361e2824446954e1e1213"
-      ]
-    }
-  }
-}
