--- scripts/examples/example_dist_generated_20.json
@@ -1,112 +1,0 @@
-{
-  "totalAmount": "44950004289467",
-  "merkleRoot": "0xb32f71b4cfc8a40d61161a3f922d56fa82529e754a3e997aebefefa429e34360",
-  "claims": {
-    "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC": {
-      "beneficiary": "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC",
-      "amount": 1500000000000,
-      "proof": [
-        "0xdd81146ab7489c6fbd5d7dca0405fca29e0ae27951e32267888ef85d6ec7e970",
-        "0xfa42dd59b8f79b2c995fb21d6b86ba5cc50aebcd465554c517a0855b19993e4e",
-        "0x3d144506d5302d0bf7f0facda134dfb6e8ad9f5592f7912520fd61761bfecb00",
-        "0xd5d0b028b9f0461f15450864b00cc762b1b87df7fad6bfe1c4f047f16463e5ba"
-      ]
-    },
-    "0x005511deD6Aa2012DaE1307290779ED52639cCee": {
-      "beneficiary": "0x005511deD6Aa2012DaE1307290779ED52639cCee",
-      "amount": 2200000000000,
-      "proof": [
-        "0x095b0fb750555ad42f0d1b8ede0dbfc9d8189e37a761b4152ed8597da8170b6b",
-        "0x5133d976adf49a59712b257f54e88b24345a38a5fa646cb69411f09d6fa5639b",
-        "0x2c659265c1496356d52f326db6fa61ebca43026b075ea5018f2c216cc4cf8beb",
-        "0xd5d0b028b9f0461f15450864b00cc762b1b87df7fad6bfe1c4f047f16463e5ba"
-      ]
-    },
-    "0x9eeE3a95f91b23f2bC186B5977E2fC9d4636a229": {
-      "beneficiary": "0x9eeE3a95f91b23f2bC186B5977E2fC9d4636a229",
-      "amount": 9370000000000,
-      "proof": [
-        "0xe0c234c4e7a7f1bf074df909169914ff97e1aa96843c5d8c8fb141d8da2a3973",
-        "0x3a58660debaf7daafeb453e9955441c40562d1695200aaf223ade75c7d486f7b",
-        "0x3d144506d5302d0bf7f0facda134dfb6e8ad9f5592f7912520fd61761bfecb00",
-        "0xd5d0b028b9f0461f15450864b00cc762b1b87df7fad6bfe1c4f047f16463e5ba"
-      ]
-    },
-    "0x61b4b98460cd58BCc3758E6885AE4aaAC7f213dE": {
-      "beneficiary": "0x61b4b98460cd58BCc3758E6885AE4aaAC7f213dE",
-      "amount": 4240000000000,
-      "proof": [
-        "0xf17e54bf7e0ef874d7719d1712573d2dccf37c9396798ca4a7f746128bdfd9b3",
-        "0xf3ff7235c1ad5a01a507294be021fc2f52bf8b562067d283ccdad6da340368de",
-        "0xf5113b9916191973c7bd775035c157bf442c1a28040089e848df45695d5a8698"
-      ]
-    },
-    "0x38445383F364BC835BDAa25e09ab8F1F18479CaE": {
-      "beneficiary": "0x38445383F364BC835BDAa25e09ab8F1F18479CaE",
-      "amount": 6660000000000,
-      "proof": [
-        "0xef21413c88555d4a0fd6045cac820bcc75e9b297a447f4898432ed36d1d7bcd9",
-        "0xf3ff7235c1ad5a01a507294be021fc2f52bf8b562067d283ccdad6da340368de",
-        "0xf5113b9916191973c7bd775035c157bf442c1a28040089e848df45695d5a8698"
-      ]
-    },
-    "0x0C51419B21c884388De6118042f0f9E5b5296A32": {
-      "beneficiary": "0x0C51419B21c884388De6118042f0f9E5b5296A32",
-      "amount": 7780000000000,
-      "proof": [
-        "0xab62cd61bcfff8c11341f9129c106f71326fb5ddbb46eb7e2133400c441166fe",
-        "0xe5b5c5b281185e22d3b9246c448a1e2036b74174a89ec572c94f00cf14ffe88c",
-        "0x2c659265c1496356d52f326db6fa61ebca43026b075ea5018f2c216cc4cf8beb",
-        "0xd5d0b028b9f0461f15450864b00cc762b1b87df7fad6bfe1c4f047f16463e5ba"
-      ]
-    },
-    "0x8324e107Ab17918Ec3FCE05dC581913223762eb5": {
-      "beneficiary": "0x8324e107Ab17918Ec3FCE05dC581913223762eb5",
-      "amount": 4350000000000,
-      "proof": [
-        "0xa2a3d00b3a11e80496a6355b69af53362b40af04a35040e66d2b88d3dfa096a0",
-        "0xe5b5c5b281185e22d3b9246c448a1e2036b74174a89ec572c94f00cf14ffe88c",
-        "0x2c659265c1496356d52f326db6fa61ebca43026b075ea5018f2c216cc4cf8beb",
-        "0xd5d0b028b9f0461f15450864b00cc762b1b87df7fad6bfe1c4f047f16463e5ba"
-      ]
-    },
-    "0x729EE12F085cf02a7c365f81192f7774572663cA": {
-      "beneficiary": "0xecd958EF8F5Dc16Cec4ef7d5441bFd9163C3f725",
-      "amount": 8850000000000,
-      "proof": [
-        "0x9639f21e83e232f03cb61bb5f9ed1bcffea092e3405da991f4c9bb6021c55667",
-        "0x5133d976adf49a59712b257f54e88b24345a38a5fa646cb69411f09d6fa5639b",
-        "0x2c659265c1496356d52f326db6fa61ebca43026b075ea5018f2c216cc4cf8beb",
-        "0xd5d0b028b9f0461f15450864b00cc762b1b87df7fad6bfe1c4f047f16463e5ba"
-      ]
-    },
-    "0xd61857D7f21244eAd9581688F0E26da419Db3518": {
-      "beneficiary": "0xC1cfb0892ee7ab0538EF519114cFd99165788c3d",
-      "amount": 78739,
-      "proof": [
-        "0xb23b10194c08332ce6cf7f45a4b8b9841050bd2db7582c95a7c3dbed2cd069fd",
-        "0xf5113b9916191973c7bd775035c157bf442c1a28040089e848df45695d5a8698"
-      ]
-    },
-    "0x6a873316dfAbe694D1b015a929f7C2Ed432e38E4": {
-      "beneficiary": "0x017ed213dfFc8cD77f2328371726CCDAE785145C",
-      "amount": 81016,
-      "proof": [
-        "0xe446a2b9cf9e01c85fb6064580df048ef2eabfb7c76245eca68540cc45443491",
-        "0x3a58660debaf7daafeb453e9955441c40562d1695200aaf223ade75c7d486f7b",
-        "0x3d144506d5302d0bf7f0facda134dfb6e8ad9f5592f7912520fd61761bfecb00",
-        "0xd5d0b028b9f0461f15450864b00cc762b1b87df7fad6bfe1c4f047f16463e5ba"
-      ]
-    },
-    "0x4492105885AA9eAa556aC6F96B0e49917bC1A080": {
-      "beneficiary": "0x044112544758015E7Af23f19c9cf9457b726650B",
-      "amount": 4129712,
-      "proof": [
-        "0xd56788d170d210f8cb40ffc969c51ee35e80a4f2d800204e9cb146aac5209a18",
-        "0xfa42dd59b8f79b2c995fb21d6b86ba5cc50aebcd465554c517a0855b19993e4e",
-        "0x3d144506d5302d0bf7f0facda134dfb6e8ad9f5592f7912520fd61761bfecb00",
-        "0xd5d0b028b9f0461f15450864b00cc762b1b87df7fad6bfe1c4f047f16463e5ba"
-      ]
-    }
-  }
-}
