--- scripts/examples/example_proof_generated_10.json
@@ -1,15 +1,0 @@
-{
-  "merkleRoot": "0x971af7943ef35c1951d2b34a0f543aceea4d990b43386fdc0fc61bfc50644c05",
-  "claims": {
-    "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC": {
-      "beneficiary": "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC",
-      "amount": 1000000000000,
-      "proof": [
-        "0x753e810ed621eaa35348ddb2bbcf81e93828fd6a69b15029589ea31c093c99ef",
-        "0x4662c3b26baece43e1444c93540566b8a9ed3e88b259b63970091be98c8fbcc8",
-        "0x07f117b1d7aa852c979367c249b57b7c1cb98dfeccb06ac0ee719d6a07880d99",
-        "0x2c284433e85395a49e74b68c571b119870574c674d7c783a7f3c74c606610502"
-      ]
-    }
-  }
-}
