--- scripts/examples/example_stake_list_10.json
@@ -1,42 +1,0 @@
-{
-  "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC": {
-    "beneficiary": "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC",
-    "amount": 1000000000000
-  },
-  "0x005511deD6Aa2012DaE1307290779ED52639cCee": {
-    "beneficiary": "0x005511deD6Aa2012DaE1307290779ED52639cCee",
-    "amount": 2000000000000
-  },
-  "0x9eeE3a95f91b23f2bC186B5977E2fC9d4636a229": {
-    "beneficiary": "0x9eeE3a95f91b23f2bC186B5977E2fC9d4636a229",
-    "amount": 9360000000000
-  },
-  "0x61b4b98460cd58BCc3758E6885AE4aaAC7f213dE": {
-    "beneficiary": "0x61b4b98460cd58BCc3758E6885AE4aaAC7f213dE",
-    "amount": 4230000000000
-  },
-  "0x38445383F364BC835BDAa25e09ab8F1F18479CaE": {
-    "beneficiary": "0x38445383F364BC835BDAa25e09ab8F1F18479CaE",
-    "amount": 6600000000000
-  },
-  "0x0C51419B21c884388De6118042f0f9E5b5296A32": {
-    "beneficiary": "0x0C51419B21c884388De6118042f0f9E5b5296A32",
-    "amount": 7770000000000
-  },
-  "0x8324e107Ab17918Ec3FCE05dC581913223762eb5": {
-    "beneficiary": "0x8324e107Ab17918Ec3FCE05dC581913223762eb5",
-    "amount": 4340000000000
-  },
-  "0x729EE12F085cf02a7c365f81192f7774572663cA": {
-    "beneficiary": "0xecd958EF8F5Dc16Cec4ef7d5441bFd9163C3f725",
-    "amount": 8840000000000
-  },
-  "0x91aa9159242c298810b280E3AfFC0b4B289b60cc": {
-    "beneficiary": "0xecd958EF8F5Dc16Cec4ef7d5441bFd9163C3f725",
-    "amount": 3000000000000
-  },
-  "0xE4c8d3bcf8C87D73CE38Ab2DC288d309072ee4E7": {
-    "beneficiary": "0xecd958EF8F5Dc16Cec4ef7d5441bFd9163C3f725",
-    "amount": 534000000000
-  }
-}
