--- scripts/examples/example_stake_list_20.json
@@ -1,46 +1,0 @@
-{
-  "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC": {
-    "beneficiary": "0x2f9e94217C3f3493A2E65676E0F409fF5642A9AC",
-    "amount": 1500000000000
-  },
-  "0x005511deD6Aa2012DaE1307290779ED52639cCee": {
-    "beneficiary": "0x005511deD6Aa2012DaE1307290779ED52639cCee",
-    "amount": 2200000000000
-  },
-  "0x9eeE3a95f91b23f2bC186B5977E2fC9d4636a229": {
-    "beneficiary": "0x9eeE3a95f91b23f2bC186B5977E2fC9d4636a229",
-    "amount": 9370000000000
-  },
-  "0x61b4b98460cd58BCc3758E6885AE4aaAC7f213dE": {
-    "beneficiary": "0x61b4b98460cd58BCc3758E6885AE4aaAC7f213dE",
-    "amount": 4240000000000
-  },
-  "0x38445383F364BC835BDAa25e09ab8F1F18479CaE": {
-    "beneficiary": "0x38445383F364BC835BDAa25e09ab8F1F18479CaE",
-    "amount": 6660000000000
-  },
-  "0x0C51419B21c884388De6118042f0f9E5b5296A32": {
-    "beneficiary": "0x0C51419B21c884388De6118042f0f9E5b5296A32",
-    "amount": 7780000000000
-  },
-  "0x8324e107Ab17918Ec3FCE05dC581913223762eb5": {
-    "beneficiary": "0x8324e107Ab17918Ec3FCE05dC581913223762eb5",
-    "amount": 4350000000000
-  },
-  "0x729EE12F085cf02a7c365f81192f7774572663cA": {
-    "beneficiary": "0xecd958EF8F5Dc16Cec4ef7d5441bFd9163C3f725",
-    "amount": 8850000000000
-  },
-  "0xd61857D7f21244eAd9581688F0E26da419Db3518": {
-    "beneficiary": "0xC1cfb0892ee7ab0538EF519114cFd99165788c3d",
-    "amount": 78739
-  },
-  "0x6a873316dfAbe694D1b015a929f7C2Ed432e38E4": {
-    "beneficiary": "0x017ed213dfFc8cD77f2328371726CCDAE785145C",
-    "amount": 81016
-  },
-  "0x4492105885AA9eAa556aC6F96B0e49917bC1A080": {
-    "beneficiary": "0x044112544758015E7Af23f19c9cf9457b726650B",
-    "amount": 4129712
-  }
-}
