+++ src/scripts/verify_merkle_dist.js
