--- src/stakingrewards/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "stakingrewards",
-  "version": "1.0.0",
-  "description": "Staking rewards calculation in Threshold Network",
-  "main": "subgraph.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "GPL-3.0-or-later",
-  "dependencies": {
-    "@urql/core": "^2.5.0",
-    "bignumber.js": "^9.0.2",
-    "ethers": "^5.6.9",
-    "graphql": "^16.5.0",
-    "isomorphic-unfetch": "^3.1.0",
-    "keccak256": "^1.0.6",
-    "merkletreejs": "^0.2.32"
-  }
-}
